https://issues.apache.org/activemq/browse/AMQCPP-265
Fix a couple of windows warnings

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@834997 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test-integration/activemq/test/ExpirationTest.cpp
@@ -19,7 +19,7 @@
 
 #include <decaf/lang/Runnable.h>
 #include <decaf/lang/Thread.h>
-#include <decaf/lang/Integer.h>
+#include <decaf/lang/Long.h>
 #include <decaf/util/UUID.h>
 
 #include <sstream>
@@ -81,7 +81,7 @@ namespace test {
                 }
 
                 // Create the Thread Id String
-                string threadIdStr = Integer::toString( Thread::getId() );
+                string threadIdStr = Long::toString( Thread::getId() );
 
                 // Create a messages
                 string text = (string)"Hello world! from thread " + threadIdStr;
